Fix device rotation (user bitmap saving/restoring and repositioning viewport)
Reorienting the device to cause a configuration change no longer crashes the
app. Additionally the user drawn bitmap is restored after the rotation
and repositioned based on the center of the viewport.

Also can switch between layouts to some extent based on screen dimensions
using a single Activity that decides which of two child Fragments to load
which each go on to loading their own child Fragments. This is opposed to
previously having one driver Activity decide on another Activity which
then load its Fragments). Due to ICS not supporting nested Fragments, the
Fragment API must be switched to the support library's implementation
which does support this on older versions.