Improve tool interface
The Tool class now executes code that was being repeated
by the subclasses such as checking the cancelled variable.
To achieve this it calls an associated abstract function
within each of the start/move/end functions for subclass's
implementation.

Command's end() function now returns a Path which
represents the final path taken by the tool at the end
of the lifecycle of that operation. This improves the
undo/redo system, as we don't commit drawing that took
place off of the bitmap. It also enables selection
tools.