Add Selection tools and Colour Dropper, update theme
Selection tools can be used to select areas and then blank them out,
they will be used for copy and pasting once layer support is added.
ToolAttributes now have the functions isMutator() and isSelectable()
to support selection.

The Colour Dropper tool selects colours from the drawing. In order
to support this, the start(), move() and end() funcions of the
Command interface now return a ToolReport which contains details
of what the tool has done. It includes the Path traversed by the
tool so far in the current operation and also the colour of the
pixel touched in the current event.

Finally, the app has been rethemed. The icons have become white
and the palette menu item now has a white border to stay
consistent with the theme. The theme will continue to be
improved upon.