Improve thumbnail scale, max/min zoom and gridline edge
The thumbnail is now scaled by the device's density.

The max zoom is now when there are at most 4 pixels along one
dimension of the screen, and min zoom is when the image becomes
width*dp by height*dp in dimension.

The right-most and bottom-most gridlines no longer flicker when
zooming.

Also fixed the zoom threshold to display the thumbnail.