Fix draw op undo bug
Pixels with an alpha of 0 were losing their RGB components
after being XOR'd. The cause was Bitmap.setPixels() storing
the colour data as premultiplied RGB (premultiplied with 0).
The fix required spliting the xor() function into two,
one that inserts the pixels back into a Bitmap using
Bitmap.setPixels() (used when undoing/redoing a draw op,
losing the RGB components in order to draw makes sense) and
another that doesn't (used when adding a draw op to be
compressed, we must retain the RGB components in this case).