Switch from ActionBar to Toolbar and add animation panel
The ActionBar has been replaced with a static top Toolbar to
take advantage of more support library features such as the
Navigation Drawer. There is now a side drawer that contains
the outline of the AnimationFragment. To achieve better UI
consistency, other panels are also likely to be moved to
side drawers in the future.