Add magic wand tool.
The magic wand tool selects regions that are similar in
colour to the pixel that the user touched. The threshold of
similarity is controlable by a slider in the Toolbax options
panel. A threshold of 0 means the colours must be identical, a
theshold of 100 means that all colours will be selected.
For example, using distance of two colours in RGB-space, the
distance between red and magenta is 50 out of 100.