Fix theming issues
Toolbar toasts and popup menu now used the proper theme.
Previously the text background was not the same as rest of the
toast's background. The theme for the load dialog has also been
fixed. The problem was that the 'windowBackground' attribute
used for the drawing area was also being set for the dialog,
making the places that should have been transparent become
grey.