Fix thumbnail size and checkerboard interface
Thumbnail maintains a constant size, regardless of size of the
bitmap that it is previewing.

The transparency checkerboard has an additional function to draw
the tiled background taking only the rectangular area to be
tiled (no need for an additional visible area rectangle). This
simplifies the code for the thumbnail to draw its checkerboard
background.